+++ molviewspec/molviewspec/builder.py
@@ -15,6 +15,8 @@ from typing_extensions import Self
 
 from molviewspec.nodes import (
     AngleMeasurementParams,
+    AnimationNode,
+    AnimationParams,
     ArrowParams,
     BoxParams,
     CameraParams,
@@ -30,14 +32,19 @@ from molviewspec.nodes import (
     ComponentFromUriParams,
     ComponentInlineParams,
     ComponentSelectorT,
+    ContinuousPalette,
     CustomT,
     DescriptionFormatT,
+    DiscretePalette,
     DistanceMeasurementParams,
     DownloadParams,
+    EasingKindT,
     EllipseParams,
     EllipsoidParams,
     FocusInlineParams,
     GlobalMetadata,
+    InterpolationKindParams,
+    InterpolationKindT,
     LabelAttachmentT,
     LabelFromSourceParams,
     LabelFromUriParams,
@@ -64,6 +71,7 @@ from molviewspec.nodes import (
     Snapshot,
     SnapshotMetadata,
     State,
+    States,
     StructureParams,
     SurfaceTypeT,
     TooltipFromSourceParams,
@@ -228,6 +236,7 @@ class _TransformMixin(_BuilderProtocol):
         self,
         *,
         rotation: Sequence[float] | None = None,
+        rotation_center: Vec3[float] | Literal["centroid"] | None = None,
         translation: Sequence[float] | None = None,
         matrix: Sequence[float] | None = None,
         custom: CustomT = None,
@@ -236,6 +245,7 @@ class _TransformMixin(_BuilderProtocol):
         """
         Transform the object by applying a rotation matrix and/or translation vector OR a full transform matrix.
         :param rotation: 9d vector describing the rotation, in column major (j * 3 + i indexing) format, this is equivalent to Fortran-order in numpy, to be multiplied from the left (default: identity matrix)
+        :param rotation_center: 3d vector (or centroid reference) describing the center of rotation (default: (0, 0, 0))
         :param translation: 3d vector describing the translation (default: (0, 0, 0))
         :param matrix: 16d (4x4) vector describing the transformation matrix, in column major (j * 4 + i indexing) format, this is equivalent to Fortran-order in numpy, to be multiplied from the left. Takes precedence over `rotation` and `translation`.
         :param custom: optional, custom data to attach to this node
@@ -265,6 +275,7 @@ class _TransformMixin(_BuilderProtocol):
         self,
         *,
         rotation: Sequence[float] | None = None,
+        rotation_center: Vec3[float] | Literal["centroid"] | None = None,
         translation: Sequence[float] | None = None,
         matrix: Sequence[float] | None = None,
         custom: CustomT = None,
@@ -273,6 +284,7 @@ class _TransformMixin(_BuilderProtocol):
         """
         Instantiate the object by applying a rotation matrix and/or translation vector OR a full transform matrix.
         :param rotation: 9d vector describing the rotation, in column major (j * 3 + i indexing) format, this is equivalent to Fortran-order in numpy, to be multiplied from the left (default: identity matrix)
+        :param rotation_center: 3d vector (or centroid reference) describing the center of rotation (default: (0, 0, 0))
         :param translation: 3d vector describing the translation (default: (0, 0, 0))
         :param matrix: 16d (4x4) vector describing the transformation matrix, in column major (j * 4 + i indexing) format, this is equivalent to Fortran-order in numpy, to be multiplied from the left. Takes precedence over `rotation` and `translation`.
         :param custom: optional, custom data to attach to this node
@@ -445,7 +457,8 @@ class Root(_Base, _PrimitivesMixin, _FocusMixin, MolstarWidgetsMixin):
             linger_duration_ms=linger_duration_ms,
             transition_duration_ms=transition_duration_ms,
         )
-        return Snapshot(root=self._node, metadata=metadata)  # TODO create deep copy of node
+        animation = self._animation.node if self._animation else None
+        return Snapshot(root=self._node, metadata=metadata, animation=animation)  # TODO create deep copy of node
 
     def get_state(
         self,
@@ -453,14 +466,12 @@ class Root(_Base, _PrimitivesMixin, _FocusMixin, MolstarWidgetsMixin):
         title: str | None = None,
         description: str | None = None,
         description_format: DescriptionFormatT | None = None,
-        indent: int | None = 2,
     ) -> State:
         """
         Return single-state MVSJ State object. Can be enriched with metadata.
         :param title: optional title of the scene
         :param description: optional detailed description of the scene
         :param description_format: format of the description
-        :param indent: control format by specifying if and how to indent attributes
         :return: JSON string that resembles that whole state
         """
         metadata = GlobalMetadata(
@@ -469,6 +480,11 @@ class Root(_Base, _PrimitivesMixin, _FocusMixin, MolstarWidgetsMixin):
             description_format=description_format,
             # `version` and `timestamp` added by the constructor
         )
+
+        if self._animation is not None:
+            snapshot = self.get_snapshot(title=title, description=description, description_format=description_format)
+            return States(snapshots=[snapshot], metadata=metadata)
+
         return State(root=self._node, metadata=metadata)
 
     def save_state(
@@ -528,7 +544,7 @@ class Root(_Base, _PrimitivesMixin, _FocusMixin, MolstarWidgetsMixin):
     def canvas(
         self,
         *,
-        background_color: ColorT,
+        background_color: ColorT | None = None,
         custom: CustomT = None,
         ref: RefT = None,
     ) -> Root:
@@ -562,6 +578,38 @@ class Root(_Base, _PrimitivesMixin, _FocusMixin, MolstarWidgetsMixin):
         node = Node(kind="download", params=params)
         self._add_child(node)
         return Download(node=node, root=self._root)
+
+    _animation: Animation | None = PrivateAttr(None)
+
+    def animation(
+        self,
+        *,
+        frame_time_ms: float | None = None,
+        duration_ms: float | None = None,
+        autoplay: bool | None = None,
+        loop: bool | None = None,
+        include_camera: bool | None = None,
+        include_canvas: bool | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Animation:
+        """
+        Create an animation node.
+        :param frame_time_ms: Duration of each animation frame in milliseconds (default 1000/60)
+        :param duration_ms: Total duration of the animation in milliseconds, if not specified, computed as maximum of all transitions
+        :param autoplay: Whether the animation should start playing automatically (default True)
+        :param loop: Whether the animation should loop (default False)
+        :param include_camera: Whether to include camera movements in the animation (default False)
+        :param include_canvas: Whether to include canvas changes in the animation (default False)
+        :param custom: Optional custom data to attach to this node.
+        :param ref: Optional reference that can be used to access this node.
+        """
+        if self._animation is not None:
+            return self._animation
+
+        params = make_params(AnimationParams, locals())
+        self._animation = Animation(params=params)
+        return self._animation
 
 
 class Download(_Base):
@@ -1488,7 +1536,7 @@ class Primitives(_Base, _FocusMixin):
         *,
         start: PrimitivePositionT,
         end: PrimitivePositionT | None = None,
-        direction: Vec3 | None = None,
+        direction: Vec3[float] | None = None,
         length: float | None = None,
         show_start_cap: bool | None = None,
         start_cap_length: float | None = None,
@@ -1648,8 +1696,8 @@ class Primitives(_Base, _FocusMixin):
         *,
         center: PrimitivePositionT,
         as_circle: bool | None = None,
-        major_axis: Vec3 | None = None,
-        minor_axis: Vec3 | None = None,
+        major_axis: Vec3[float] | None = None,
+        minor_axis: Vec3[float] | None = None,
         major_axis_endpoint: PrimitivePositionT | None = None,
         minor_axis_endpoint: PrimitivePositionT | None = None,
         radius_major: float | None = None,
@@ -1693,12 +1741,12 @@ class Primitives(_Base, _FocusMixin):
         self,
         *,
         center: PrimitivePositionT,
-        major_axis: Vec3 | None = None,
-        minor_axis: Vec3 | None = None,
+        major_axis: Vec3[float] | None = None,
+        minor_axis: Vec3[float] | None = None,
         major_axis_endpoint: PrimitivePositionT | None = None,
         minor_axis_endpoint: PrimitivePositionT | None = None,
-        radius: Vec3 | float | None = None,
-        radius_extent: Vec3 | float | None = None,
+        radius: Vec3[float] | float | None = None,
+        radius_extent: Vec3[float] | float | None = None,
         color: ColorT | None = None,
         tooltip: str | None = None,
         custom: CustomT = None,
@@ -1755,7 +1803,7 @@ class Primitives(_Base, _FocusMixin):
         self,
         *,
         center: PrimitivePositionT,
-        extent: Vec3 | None = None,
+        extent: Vec3[float] | None = None,
         show_faces: bool = True,
         face_color: ColorT | None = None,
         show_edges: bool = False,
@@ -1787,5 +1835,243 @@ class Primitives(_Base, _FocusMixin):
         """
         params = make_params(BoxParams, {"kind": "box", **locals()})
         node = Node(kind="primitive", params=params)
+        self._add_child(node)
+        return self
+
+
+class Animation:
+    def __init__(self, params: AnimationParams) -> None:
+        self.node = AnimationNode(kind="animation", params=params)
+
+    def _add_child(self, node: Node) -> None:
+        if self.node.children is None:
+            self.node.children = []
+        self.node.children.append(node)
+
+    @overload
+    def interpolate(
+        self,
+        *,
+        kind: Literal["scalar"],
+        target_ref: str,
+        property: str | list[str | int],
+        start_ms: float | None = None,
+        duration_ms: float,
+        easing: EasingKindT | None = None,
+        noise_magnitude: float | None = None,
+        start: float | Sequence[float] | None = None,
+        end: float | Sequence[float] | None = None,
+        frequency: int | None = None,
+        alternate_direction: bool | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Animation:
+        """
+        Scalar interpolation
+        :param kind: the kind of interpolation to use
+        :param target_ref: reference to the node that should be animated.
+        :param property: value accessor.
+        :param start_ms: start time of the transition in milliseconds (default: 0)
+        :param duration_ms: duration of the transition in milliseconds.
+        :param easing: easing function to use for the transition (default: linear)
+        :param noise_magnitude: amount of noise to add to the interpolation (default: 0)
+        :param start: start value of the interpolation, if a list of values is provided, each element will be interpolated separately, (default: parent state value)
+        :param end: end value of the interpolation, if a list of values is provided, each element will be interpolated separately, if unset, only noise is applied.
+        :param frequency: determines how many times the interpolation loops. Current T = frequency * t mod 1.
+        :param alternate_direction: whether to alternate the direction of the interpolation for frequency > 1.
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: this builder
+        """
+        ...
+
+    @overload
+    def interpolate(
+        self,
+        *,
+        kind: Literal["vec3"],
+        target_ref: str,
+        property: str | list[str | int],
+        start_ms: float | None = None,
+        duration_ms: float,
+        easing: EasingKindT | None = None,
+        noise_magnitude: float | None = None,
+        start: Sequence[float] | None = None,
+        end: Sequence[float] | None = None,
+        spherical: bool | None = None,
+        frequency: int | None = None,
+        alternate_direction: bool | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Animation:
+        """
+        Vector3 interpolation
+        :param kind: the kind of interpolation to use
+        :param target_ref: reference to the node that should be animated.
+        :param property: value accessor.
+        :param start_ms: start time of the transition in milliseconds (default: 0)
+        :param duration_ms: duration of the transition in milliseconds.
+        :param easing: easing function to use for the transition (default: linear)
+        :param noise_magnitude: amount of noise to add to the interpolation (default: 0)
+        :param start: start value of the interpolation (default: parent state value)
+        :param end: end value of the interpolation, if unset, only noise is applied.
+        :param spherical: whether to use spherical interpolation (default: False)
+        :param frequency: determines how many times the interpolation loops. Current T = frequency * t mod 1.
+        :param alternate_direction: whether to alternate the direction of the interpolation for frequency > 1.
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: this builder
+        """
+        ...
+
+    @overload
+    def interpolate(
+        self,
+        *,
+        kind: Literal["rotation_matrix"],
+        target_ref: str,
+        property: str | list[str | int],
+        start_ms: float | None = None,
+        duration_ms: float,
+        easing: EasingKindT | None = None,
+        noise_magnitude: float | None = None,
+        start: Sequence[float] | None = None,
+        end: Sequence[float] | None = None,
+        frequency: int | None = None,
+        alternate_direction: bool | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Animation:
+        """
+        Vector3 interpolation
+        :param kind: the kind of interpolation to use
+        :param target_ref: reference to the node that should be animated.
+        :param property: value accessor.
+        :param start_ms: start time of the transition in milliseconds (default: 0)
+        :param duration_ms: duration of the transition in milliseconds.
+        :param easing: easing function to use for the transition (default: linear)
+        :param noise_magnitude: amount of noise to add to the interpolation (default: 0)
+        :param start: start value of the interpolation (default: parent state value)
+        :param end: end value of the interpolation, if unset, only noise is applied.
+        :param frequency: determines how many times the interpolation loops. Current T = frequency * t mod 1.
+        :param alternate_direction: whether to alternate the direction of the interpolation for frequency > 1.
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: this builder
+        """
+        ...
+
+    @overload
+    def interpolate(
+        self,
+        *,
+        kind: Literal["transform_matrix"],
+        target_ref: str,
+        property: str | list[str | int],
+        start_ms: float | None = None,
+        duration_ms: float,
+        pivot: Vec3[float] | None = None,
+        rotation_start: Mat3 | None = None,
+        rotation_end: Mat3 | None = None,
+        rotation_noise_magnitude: float | None = None,
+        rotation_easing: EasingKindT | None = None,
+        rotation_frequency: int | None = None,
+        rotation_alternate_direction: bool | None = None,
+        translation_start: Vec3[float] | None = None,
+        translation_end: Vec3[float] | None = None,
+        translation_noise_magnitude: float | None = None,
+        translation_easing: EasingKindT | None = None,
+        translation_frequency: int | None = None,
+        translation_alternate_direction: bool | None = None,
+        scale_start: Vec3[float] | None = None,
+        scale_end: Vec3[float] | None = None,
+        scale_noise_magnitude: float | None = None,
+        scale_easing: EasingKindT | None = None,
+        scale_frequency: int | None = None,
+        scale_alternate_direction: bool | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Animation:
+        """
+        Vector3 interpolation
+        :param kind: the kind of interpolation to use
+        :param target_ref: reference to the node that should be animated.
+        :param property: value accessor.
+        :param start_ms: start time of the transition in milliseconds (default: 0)
+        :param duration_ms: duration of the transition in milliseconds.
+        :param pivot: optional, pivot point for the rotation and scale transformation
+        :param rotation_start: optional, starting rotation matrix
+        :param rotation_end: optional, ending rotation matrix, if unset, only noise is applied.
+        :param rotation_noise_magnitude: optional, magnitude of the noise to apply to the rotation
+        :param rotation_easing: optional, easing function to use for the rotation
+        :param rotation_frequency: optional, frequency of the rotation animation
+        :param rotation_alternate_direction: optional, whether to alternate the direction of the rotation animation
+        :param translation_start: optional, starting translation vector
+        :param translation_end: optional, ending translation vector, if unset, only noise is applied.
+        :param translation_noise_magnitude: optional, magnitude of the noise to apply to the translation
+        :param translation_easing: optional, easing function to use for the translation
+        :param translation_frequency: optional, frequency of the translation animation
+        :param translation_alternate_direction: optional, whether to alternate the direction of the translation animation
+        :param scale_start: optional, starting scale vector
+        :param scale_end: optional, ending scale vector, if unset, only noise is applied.
+        :param scale_noise_magnitude: optional, magnitude of the noise to apply to the scale
+        :param scale_easing: optional, easing function to use for the scale
+        :param scale_frequency: optional, frequency of the scale animation
+        :param scale_alternate_direction: optional, whether to alternate the direction of the scale animation
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: this builder
+        """
+        ...
+
+    @overload
+    def interpolate(
+        self,
+        *,
+        kind: Literal["color"],
+        target_ref: str,
+        property: str | list[str | int],
+        start_ms: float | None = None,
+        duration_ms: float,
+        start: ColorT | None = None,
+        end: ColorT | None = None,
+        palette: ContinuousPalette | DiscretePalette | None = None,
+        frequency: int | None = None,
+        alternate_direction: bool | None = None,
+        custom: CustomT = None,
+        ref: RefT = None,
+    ) -> Animation:
+        """
+        Color interpolation
+        :param kind: the kind of interpolation to use
+        :param target_ref: reference to the node that should be animated.
+        :param property: value accessor.
+        :param start_ms: start time of the transition in milliseconds (default: 0)
+        :param duration_ms: duration of the transition in milliseconds.
+        :param start: starting color (default: parent state value)
+        :param end: ending color
+        :param palette: color palette to use for the interpolation, if set overrides start and end colors.
+        :param frequency: determines how many times the interpolation loops. Current T = frequency * t mod 1.
+        :param alternate_direction: whether to alternate the direction of the interpolation for frequency > 1.
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :return: this builder
+        """
+        ...
+
+    def interpolate(
+        self, *, kind: InterpolationKindT, custom: CustomT = None, ref: RefT = None, **kwargs: Any
+    ) -> Animation:
+        """
+        Add a representation for this component.
+        :param kind: the kind of interpolation to use
+        :param custom: optional, custom data to attach to this node
+        :param ref: optional, reference that can be used to access this node
+        :param kwargs: optional, node-specific params
+        :return: this builder
+        """
+        params_class = InterpolationKindParams.get(kind)
+        params = make_params(params_class, locals(), **kwargs)  # type: ignore
+        node = AnimationNode(kind="interpolate", params=params)
         self._add_child(node)
         return self