+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file, following t
 
 ## [Unreleased]
 
+- Add snapshot animation support
 - Add `load/s` to `MVSJ` object
 - Add `find_ref` to `State` and `MVSJ` objects
 - Add `ref/custom` props to builder methods where previously missing
@@ -17,7 +18,9 @@ All notable changes to this project will be documented in this file, following t
 - Add `matrix` field support to `TransformParams`
 - Add `instance` node type
 - Add `surface_type` (`molecular` / `gaussian`) to the surface representation
+- Add `transform.rotation_center` to support rotating objects around their centroids or defined points
 - Support transforms and instancing on structures, components, and volumes
+- Make `canvas.background_color` optional
 
 ## [v1.6.0] - 2025-04-22